backend/requirements.txt
# Multi-View RAG Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
neo4j>=5.15.0

# AI/ML
openai>=1.10.0
numpy>=1.24.0
scipy>=1.11.0
spacy>=3.5.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Utilities
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
